MaximumDraws0.py
A person is getting ready to leave and needs a pair of matching socks. If there are n colors of socks in the drawer, how many socks need to be removed to be certain of having a matching pair?

Example
n = 2
There are 2 colors of socks in the drawer. If they remove 2 socks, they may not match. The minimum number to insure success is 3.

Function Description
- Complete the maximumDraws function in the editor below.

maximumDraws has the following parameter:
- int n: the number of colors of socks

Returns
- int: the minimum number of socks to remove to guarantee a matching pair.

Input Format
The first line contains the number of test cases, t.
Each of the following t lines contains an integer n.

Constraints
- 1 <= t <= 1000
- 0 < n < 10^6