Create ComboMeal0.py
A fast-food chain menu is selling a burger, a can of soda, and a combo meal containing a burger and a can of soda, at prices known to you.

They have chosen the selling price for each item by first determining the total cost of making the individual items and then adding a fixed value to it, representing their profit. Assume that the cost of making a regular burger is fixed and the cost of making a regular soda is fixed.

For example, if the cost of making a regular burger is 206, the cost of making a regular soda is 145 and the fixed profit is 69, then the fast-food chain will set selling prices as:

Given the price of a burger, a can of soda and a combo meal on the menu, your task is to compute the fixed profit.

Complete the function named profit which takes in three integers denoting selling price of a burger, a can of soda and a combo meal respectively, and returns an integer denoting the fixed profit.

Input Format
The first line contains t, the number of scenarios. The following lines describe the scenarios.

Each scenario is described by a single line containing three space-separated integers, b, s, and c, denoting how much a burger, a can of soda and a combo meal cost respectively.

Constraints
- 1 ≤ t ≤ 100
- 3 ≤ c ≤ 2000
- 2 ≤ b, s < c