Create Foundations1MatrixAddition0.py
Matrix Addition
Add the two 3 × 3 matrices given below and find the integers corresponding to a, b, c, d, e, f, g, h, and i:

                    [1  2  3]        [4  5  6]       [a b c]
                    [2  3  4]   +   [7  8  9]  =  [d e f] 
                    [1  1  1]        [4  5  7]       [g h i]

To submit your answer, enter the resultant values of each of the nine integers (i.e., a, b, c, d, e, f, g, h, and i) on a new line and click Submit Code.

Input Format
There is no input for this challenge; calculate the values of a through i using the matrices given above.

Output Format
In the text box below, enter the values of each of the nine integers on a new line. You must have a total of nine lines of output and the integers must be printed in order (i.e., a, b, c, d, e, f, g, h, and i, respectively).