Create Foundations8SystemsOfEquations0.py
There are more than one integer values of α for which the following system of equations in x, y, z has no solutions:

                                    αx + y + 2 z = 0
                                      x + 2 y + z = b
                                   2 x + y + α z = 0

What is the smallest integer value of α for which the above system has no solutions? Fill in the required integer value into the text box. Do not leave any leading or trailing spaces.