Create NumberGroups0.py
The positive odd numbers are sorted in ascending order as 1, 3, 5, 7,9, 11, 13, 15, 17, 19, ... and grouped as (1), (3,5), (7,9,11), (13, 15, 17, 19), ... and so on.

Thus, the first group is (1), the second group is (3,5), the third group is (7,9,11), etc. In general, the kᵗʰ group contains the next k elements of the sequence.

Given k, find the sum of the elements of the kᵗʰ group. For k = 3 example, for , the answer is 27:

image [https://s3.amazonaws.com/hr-assets/0/1511935621-d85a3653c7-Numbergroups-2.png]

Complete the function sumOfGroup with input integer . Return the sum of the elements of the k-th group.

Constraints
• 1 ≤ k ≤ 10⁶

Subtasks
• For 50% of the maximum score, k ≤ 10³

Sample Input
  k = 3

Sample Output
  27

Explanation
  We have k = 3. The 3rd group is (7,9,11)  and the sum of its elements is 7 + 9 + 11 = 27.